+++ husky_components/label/main.js
@@ -95,7 +95,7 @@ define(function() {
      */
     templates = {
         basic: ['<div class="' + constants.textClass + '">',
-                '   <strong><%= title %></strong>',
+                '   <% if (!!title) { %><strong><%= title %></strong><% } %>',
                 '   <span><%= description %></span>',
                 '   <div class="' + constants.counterClass + '"><span><%= counter %></span></div>',
                 '</div>'].join(''),